removed old todo, added new one
old todo is outdated: The virtual machine will allocate registers on the
fly, but the code has to be optimized for registers.

So what the code SHOULD NOT DO is work only with high-numbered
registers. It should use as low register numbers as possible, always!

new todo:
the overflow-value is not set into the overflow register. If there is a
overflow, the value which doesn't fit, will be set into the overflow register.
This has to be done.