Started to reorder things
Started to reorder the things.
There are several directories now:
- kernel/
  all VPU related stuff goes here, in subdir opcode/ there are the
  opcode-c-files

- util
  just for utils as error-stuff and debugging stuff

- reader
  for the binary-reader stuff, the thing that reads the binary file, you
  know!

This commit also adapts a lot of function prototypes and corrects
design-stuff! It is just big and really, really ugly, but if I do it in
more than just one commit, I'll never get ready!

Fixing the whole stuff will take hours, but I'll do a tag for the "new
first commit", so in the commit history there will be a mark, where to
start reading the source or so!