heapnode-realloc logic applied
- added configuration flag "--fast". If set, the VPU tries to be as fast
  as possible and does not care about how much memory is used
- removed old static function prototype
- added heapnode-ptr allocating algorithm
- shortened code at some random positions

Heapnodes get reallocated as fast as possible, with as less as possible
calls of realloc() until a limit is reached.

This limits are defined via macros, not via configuration. I chose some
simple and low limits for this first implementation, as I can't expect
how much heapnode will be allocated in production:

HEAPNODES_DOUBLE_REALLOC_LIMIT      500

This means, until 500 Heapnodes are allocated, the VPU will allocate the
number of heapnodes * 2

HEAPNODES_FAST_REALLOC_LIMIT        1000
HEAPNODES_REALLOC_FAST_STEP         50

Means, until 1000 heapnodes are allocated, the VPU will allocate 50 heapnodes.

HEAPNODES_REALLOC_STEP              5

Later on, there will be 5 heapnodes allocated if create_new_heapnode()
is called.

The create_new_heapnode() returns a pointer to the first new heapnode,
so the calling function can use it imidiately and does not have to make
a lookup for it!