implemented module-call functions
and added a memory-helper to pass to the modules, which contains the
size of the memory-data. The actually passed memory is more, but
initialized to 0x00.

It is more, because always do a realloc() call costs time, and time is
speed. Just increase the size of the memory-helper, but tell the module
only the size of the data.