added implementation for jump-opcodes
This commit implements jump opcodes with addresses from registers and a
opcode to load the next opcode into akku. This makes sense, because the
current opcode address can be read from registers[0] easily. But if a
program wants to jump to a specific opcode (dynamically), it can load
the next-opcode-address into akku and does not have to make some maths
on the current program pointer and so on.

Opcodes implemented by this commit:

- exiti
	exit with value (constant in bytecode) not with value from register

- ifzjmpra
	if akku is zero, jump to address from register, else to address

- ifzjmprr
	if akku is zero, jump to address from register, else to address from
	other register

- jmpizr
	if register is zero, jump to address from register

- jmpnzr
	if register is not zero, jump to address from register

- jmpr
	jump to address from register

- ldaddr
	load address of next opcode into akku