reformatted disassemble output
Parameters are listed in same column now, as the string representation
of a opcode is fit into a size of 10 characters. If it does not fit these 10
characters, it will still disturbe the output, but this does not matter.
Opcodes with length greater than 10 do not exist currently.

Sadly, the longest opcode string-representation size could not be calculated
that easy.