changed iteration condition
If the iteration is done until a param is zero OR the max-parameter
count is hit, the iteration will go over all MAX_PARAMETER_COUNT items.
If the OR is replaced with an AND, the iteration stops after the last
parameter, so the minx_binary_get_at() function is not called and this
speeds up the whole thing a lot!